requirements.txt
tqdm~=4.67.1
pyfiglet~=1.0.2
colorama~=0.4.6
dnspython~=2.7.0
requests~=2.32.3
python-nmap~=0.7.1
beautifulsoup4~=4.13.3